UPSTREAM: drm/client: remove the exporting of drm_client_close
The function drm_client_close is declared as static and marked as
EXPORT_SYMBOL. It's a bit confusing for an internal function to be
exported. The area of visibility for such function is its .c file
and all other modules. Other *.c files of the same module can't use it,
despite all other modules can. Relying on the fact that this is the
internal function and it's not a crucial part of the API, the patch
removes the EXPORT_SYMBOL marking of drm_client_close.

Bug: 145210207
Change-Id: I921ecf8aa1d301b02549ad54ded0bf8460a3a96a
Signed-off-by: Denis Efremov <efremov@linux.com>
Reviewed-by: Emil Velikov <emil.velikov@collabora.com>
Signed-off-by: Noralf Trønnes <noralf@tronnes.org>
Link: https://patchwork.freedesktop.org/patch/msgid/20190703170150.32548-1-efremov@linux.com
(cherry picked from commit d006a9b63d4bbb3eb1e2e616712d45acfcda7042)
Signed-off-by: Sami Tolvanen <samitolvanen@google.com>
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>