clk: bcm2835: Fix clk_register callbacks argument types
When CONFIG_CFI is enabled, there are several errors around the
register functions because bcm2835_clk_register is defined to have a
second parameter of type 'const void *' but all of the functions that
are actually called have differing types.

_REGISTER casts these functions to bcm2835_clk_register, which does not
work. Instead, add a local data variable for the actual type and make
the second parameter 'const void *' for all of these functions.

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>